Restructuring code so that the CHERAB cython dependencies are optional (fixes #10).
