+++ dev/build.sh
@@ -3,4 +3,4 @@
 CORES=`nproc --all`
 
 echo "Rebuilding VITA extension modules (in place)..."
-python setup.py build_ext -j$CORES --inplace $1 $2 $3 $4 $5
+python setup.py build_ext --include-cython -j$CORES --inplace $1 $2 $3 $4 $5
