+++ tests/fieldlines/st40_with_mesh.py
@@ -5,7 +5,7 @@ from mayavi import mlab
 from raysect.core import Point3D, World
 from raysect.primitive import import_ply
 
-from vita.core import FieldlineTracer, RK2
+from vita.modules.cherab import FieldlineTracer, RK2
 from vita.modules.fiesta import Fiesta
