--- vita/core/__init__.py
@@ -1,2 +1,0 @@
-
-from .magnetic_field import *
