--- vita/core/magnetic_field/__init__.py
@@ -1,4 +1,0 @@
-
-
-from .b_field import MagneticField
-from .tracer import FieldlineTracer, Euler, RK2
